+++ src/parser/enrichers/class/artificer/ExperimentalElixir.mjs
@@ -1,5 +1,6 @@
 /* eslint-disable class-methods-use-this */
-import { CompendiumHelper, utils } from "../../../../lib/_module.mjs";
+import { SETTINGS } from "../../../../config/_module.mjs";
+import { DDBCompendiumFolders, DDBItemImporter, utils } from "../../../../lib/_module.mjs";
 import DDBEnricherData from "../../data/DDBEnricherData.mjs";
 
 export default class ExperimentalElixir extends DDBEnricherData {
@@ -11,6 +12,8 @@ export default class ExperimentalElixir extends DDBEnricherData {
   get activity() {
     if (!this.isAction) return {};
     return {
+      name: "Roll for Experimental Elixir",
+      noConsumeTargets: true,
       activationType: "action",
       data: {
         roll: {
@@ -38,13 +41,17 @@ export default class ExperimentalElixir extends DDBEnricherData {
     };
   }
 
+  _experimentalElixirDetails = null;
+
   get experimentalElixirDetails() {
+    if (this._experimentalElixirDetails) return this._experimentalElixirDetails;
+
     const dom = utils.htmlToDocumentFragment(this.ddbParser?.ddbDefinition?.description ?? "");
 
     // Find all rows in the tbody
     const rows = dom.querySelectorAll('tbody tr');
 
-    const effects = [];
+    const details = [];
 
     rows.forEach((row) => {
       const cells = row.querySelectorAll('td');
@@ -55,27 +62,29 @@ export default class ExperimentalElixir extends DDBEnricherData {
       const strongTag = effectCell.querySelector('strong');
       if (!strongTag) return;
 
-      const title = strongTag.textContent.trim();
+      const name = strongTag.textContent.replace(".", "").trim();
       const fullText = effectCell.textContent.trim();
       const description = fullText
-        .replace(title, '')
+        .replace(name, '')
         .replace(/^[.\s]*/, '')
         .trim();
 
-      effects.push({
+      details.push({
         roll: rollValue,
-        title,
+        name,
         description,
       });
     });
 
-    return effects;
+    this._experimentalElixirDetails = details;
+
+    return details;
   }
 
   getSkeletonItem(row) {
     const itemName = `Experimental Elixir: ${row.name}`;
     return {
-      "id": utils.namedIDStub(itemName, {
+      "_id": utils.namedIDStub(itemName, {
         postfix: row.roll,
         prefix: "ee",
       }),
@@ -106,76 +115,228 @@ export default class ExperimentalElixir extends DDBEnricherData {
         },
       },
       "effects": [],
-      "flags": {},
+      "flags": {
+        "ddbimporter": {
+          "is2014": this.is2014,
+          "is2024": this.is2024,
+          "subClass": "Alchemist",
+          "class": "Artificer",
+          "experimentalElixir": true,
+        },
+      },
     };
   }
 
-  async buildItem(row) {
-    const itemData = this.getSkeletonItem(row);
-    let type = "utility";
-    let options = {
-      generateActivation: true,
-      generateConsumption: true,
-      consumeItem: true,
-      activationOverride: {
-        type: "bonus",
-        value: 1,
-        condition: "",
+  get activityMap() {
+    return this.is2014
+      ? {
+        "Healing": [
+          { number: 2, denomination: 4, bonus: "@abilities.int.mod", min: null, max: null },
+        ],
+        "Swiftness": [
+          { bonus: "10", min: null, max: null },
+        ],
+        "Resilience": [
+          { duration: "10", min: null, max: null },
+        ],
+        "Boldness": [
+          { duration: "10", min: null, max: null },
+        ],
+        "Flight": [
+          { bonus: "10", min: null, max: null },
+        ],
+        "Transformation": [
+          { duration: "10", min: null, max: null },
+        ],
+      }
+      : {
+        "Healing": [
+          { number: 2, denomination: 8, bonus: "@abilities.int.mod", min: null, max: 8 },
+          { number: 3, denomination: 8, bonus: "@abilities.int.mod", min: 9, max: 14 },
+          { number: 4, denomination: 8, bonus: "@abilities.int.mod", min: 15, max: null },
+        ],
+        "Swiftness": [
+          { bonus: "10", min: null, max: 8 },
+          { bonus: "15", min: 9, max: 14 },
+          { bonus: "20", min: 15, max: null },
+        ],
+        "Resilience": [
+          { duration: "10", min: null, max: 8 },
+          { duration: "600", min: 9, max: 14 },
+          { duration: "3600", min: 15, max: null },
+        ],
+        "Boldness": [
+          { duration: "10", min: null, max: 8 },
+          { duration: "600", min: 9, max: 14 },
+          { duration: "3600", min: 15, max: null },
+        ],
+        "Flight": [
+          { bonus: "10", min: null, max: 8 },
+          { bonus: "20", min: 9, max: 14 },
+          { bonus: "30", min: 15, max: null },
+        ],
+      };
+  }
+
+  enchantChangeMap(name) {
+    const data = this.experimentalElixirDetails.find((e) => e.name === name);
+    const result = [
+      DDBEnricherData.ChangeHelper.overrideChange(`${name} Elixir`, 20, "name"),
+      DDBEnricherData.ChangeHelper.overrideChange("potion", 20, "system.type.value"),
+      DDBEnricherData.ChangeHelper.overrideChange("1", 20, "system.uses.max"),
+      DDBEnricherData.ChangeHelper.overrideChange("charges", 20, "system.uses.per"),
+      DDBEnricherData.ChangeHelper.overrideChange("true", 20, "system.uses.prompt"),
+      DDBEnricherData.ChangeHelper.overrideChange("true", 20, "system.uses.autoDestroy"),
+      DDBEnricherData.ChangeHelper.overrideChange(data.description, 20, "system.description.value"),
+      DDBEnricherData.ChangeHelper.overrideChange("1", 20, "system.activation.cost"),
+      DDBEnricherData.ChangeHelper.overrideChange((this.is2014 ? "action" : "bonus"), 20, "system.activation.type"),
+    ];
+
+    switch (name) {
+      case "Healing":
+        result.push(
+          DDBEnricherData.ChangeHelper.overrideChange("icons/consumables/potions/bottle-round-label-cork-red.webp", 20, "img"),
+        );
+        break;
+      case "Swiftness":
+        result.push(
+          DDBEnricherData.ChangeHelper.overrideChange("icons/consumables/potions/bottle-round-label-cork-green.webp", 20, "img"),
+          DDBEnricherData.ChangeHelper.overrideChange("60", 20, "system.duration.value"),
+          DDBEnricherData.ChangeHelper.overrideChange("minute", 20, "system.duration.units"),
+        );
+        break;
+      case "Resilience":
+        result.push(
+          DDBEnricherData.ChangeHelper.overrideChange("icons/consumables/potions/bottle-round-label-cork-purple.webp", 20, "img"),
+          DDBEnricherData.ChangeHelper.overrideChange("10", 20, "system.duration.value"),
+          DDBEnricherData.ChangeHelper.overrideChange("minute", 20, "system.duration.units"),
+        );
+        break;
+      case "Boldness":
+        result.push(
+          DDBEnricherData.ChangeHelper.overrideChange("icons/consumables/potions/bottle-round-label-cork-yellow.webp", 20, "img"),
+          DDBEnricherData.ChangeHelper.overrideChange("1", 20, "system.duration.value"),
+          DDBEnricherData.ChangeHelper.overrideChange("minute", 20, "system.duration.units"),
+        );
+        break;
+      case "Flight":
+        result.push(
+          DDBEnricherData.ChangeHelper.overrideChange("icons/consumables/potions/bottle-ornate-bat-teal.webp", 20, "img"),
+          DDBEnricherData.ChangeHelper.overrideChange("10", 20, "system.duration.value"),
+          DDBEnricherData.ChangeHelper.overrideChange("minute", 20, "system.duration.units"),
+        );
+        break;
+      case "Transformation":
+        result.push(
+          DDBEnricherData.ChangeHelper.overrideChange("icons/consumables/potions/bottle-round-label-cork-blue.webp", 20, "img"),
+        );
+        break;
+    // no default
+    }
+
+    return result;
+  }
+
+  generateElixirAdditionalActivity(name) {
+    const results = [];
+    const result = {
+      constructor: {
+        name: `Use ${name}`,
+        type: "utility",
+      },
+      build: {
+        generateActivation: true,
+        generateConsumption: true,
+        consumeItem: true,
+        activationOverride: {
+          type: "bonus",
+          value: 1,
+          condition: "",
+        },
+      },
+      overrides: {
+        id: utils.namedIDStub(name, {
+          postfix: 1,
+          prefix: "eea",
+        }),
       },
     };
-    if (row.name === "Healing") {
-      type = "heal";
-      options.generateHealing = true;
-      options.healingPart = {
-        number: 2,
-        denomination: 4,
-        bonus: "@abilities.int.mod",
-        types: ["healing"],
-      };
-    } else if (row.name === "Swiftness") {
-      options.generateDuration = true;
-      options.durationOverride = {
+
+    if (name === "Healing") {
+      this.activityMap[name].forEach((a, i) => {
+        const temp = foundry.utils.deepClone(result);
+        temp.constructor.type = "heal";
+        temp.build.generateHealing = true;
+        temp.build.healingPart = {
+          number: a.number,
+          denomination: a.denomination,
+          bonus: a.bonus,
+          types: ["healing"],
+        };
+        temp.overrides = {
+          data: {
+            id: utils.namedIDStub(name, {
+              postfix: i,
+              prefix: "eea",
+            }),
+            visibility: {
+              "level": {
+                "min": a.min,
+                "max": a.max,
+              },
+              "requireAttunement": false,
+              "requireIdentification": false,
+              "requireMagic": false,
+              "identifier": "artificer",
+            },
+          },
+        };
+        results.push(temp);
+      });
+    } else if (name === "Swiftness") {
+      result.build.generateDuration = true;
+      result.build.durationOverride = {
         units: "hour",
         value: 1,
       };
-    } else if (row.name === "Resilience") {
-      options.generateDuration = true;
-      options.durationOverride = {
+      results.push(result);
+    } else if (name === "Resilience") {
+      result.build.generateDuration = true;
+      result.build.durationOverride = {
         units: "minute",
         value: 10,
       };
-    } else if (row.name === "Boldness") {
-      options.generateRoll = true;
-      options.roll = {
+      results.push(result);
+    } else if (name === "Boldness") {
+      result.build.generateRoll = true;
+      result.build.roll = {
         formula: "1d4",
         name: "Boldness Roll",
       };
-      options.generateDuration = true;
-      options.durationOverride = {
+      result.build.generateDuration = true;
+      result.build.durationOverride = {
         units: "minute",
         value: 1,
       };
-    } else if (row.name === "Flight") {
-      options.generateDuration = true;
-      options.durationOverride = {
+      results.push(result);
+    } else if (name === "Flight") {
+      result.build.generateDuration = true;
+      result.build.durationOverride = {
         units: "minute",
         value: 10,
       };
-    } else if (row.name === "Transformation") {
-      type = "cast";
-      options.generateDuration = true;
-      options.durationOverride = {
+      results.push(result);
+    } else if (name === "Transformation") {
+      result.constructor.type = "cast";
+      result.build.generateDuration = true;
+      result.build.durationOverride = {
         units: "minute",
         value: 10,
       };
-      options.generateSpell = true;
-
-      const compcompendiumSpellsIndex = CompendiumHelper.retrieveCompendiumSpellReferences(["Alter Self"], {
-        use2024Spells: this.is2024,
-      });
-
-      options.spellOverride = {
-        uuid: compcompendiumSpellsIndex[0].uuid ?? "",
+      result.build.generateSpell = true;
+      result.overrides.addSpellUuid = "Alter Self";
+      result.build.spellOverride = {
+        uuid: "",
         properties: [],
         level: null,
         challenge: {
@@ -185,36 +346,349 @@ export default class ExperimentalElixir extends DDBEnricherData {
         },
         spellbook: false,
       };
+      results.push(result);
     }
 
-    const basicActivityGenerator = new DDBImporter.lib.Enrichers.mixins.DDBBasicActivity({
-      actor: this.ddbEnricher.actor,
-      foundryFeature: itemData,
-      type,
+    return results;
+  }
+
+  generateElixirEffect(name) {
+    const effects = [];
+    if (name === "Swiftness") {
+      effects.push(...this.activityMap[name].map((data) => {
+        return {
+          name: `Experimental Elixir: ${name}`,
+          activityMatch: `Use ${name}`,
+          type: "item",
+          options: {
+            transfer: false,
+          },
+          changes: [
+            DDBEnricherData.ChangeHelper.addChange(data.bonus, 20, "system.attributes.movement.walk"),
+          ],
+          data: {
+            "_id": utils.namedIDStub(name, {
+              postfix: data.bonus,
+              prefix: "ef",
+            }),
+            "flags.ddbimporter.effectIdLevel": {
+              min: data.min,
+              max: data.max,
+            },
+          },
+        };
+      }));
+    } else if (name === "Resilience") {
+      effects.push(...this.activityMap[name].map((data) => {
+        return {
+          name: `Experimental Elixir: ${name}`,
+          activityMatch: `Use ${name}`,
+          type: "item",
+          options: {
+            transfer: false,
+            durationSeconds: data.duration,
+          },
+          changes: [
+            DDBEnricherData.ChangeHelper.addChange("1", 20, "system.attributes.ac.bonus"),
+          ],
+          data: {
+            "_id": utils.namedIDStub(name, {
+              postfix: data.bonus,
+              prefix: "ef",
+            }),
+            "flags.ddbimporter.effectIdLevel": {
+              min: data.min,
+              max: data.max,
+            },
+          },
+        };
+      }));
+    } else if (name === "Boldness") {
+      effects.push(...this.activityMap[name].map((data) => {
+        return {
+          name: `Experimental Elixir: ${name}`,
+          activityMatch: `Use ${name}`,
+          type: "item",
+          options: {
+            transfer: false,
+            durationSeconds: data.duration,
+          },
+          changes: [
+            DDBEnricherData.ChangeHelper.addChange("1d4", 20, "system.bonuses.abilities.save"),
+            DDBEnricherData.ChangeHelper.addChange("1d4", 20, "system.bonuses.msak.attack"),
+            DDBEnricherData.ChangeHelper.addChange("1d4", 20, "system.bonuses.mwak.attack"),
+            DDBEnricherData.ChangeHelper.addChange("1d4", 20, "system.bonuses.rsak.attack"),
+            DDBEnricherData.ChangeHelper.addChange("1d4", 20, "system.bonuses.rwak.attack"),
+          ],
+          data: {
+            "_id": utils.namedIDStub(name, {
+              postfix: data.bonus,
+              prefix: "ef",
+            }),
+            "flags.ddbimporter.effectIdLevel": {
+              min: data.min,
+              max: data.max,
+            },
+          },
+        };
+      }));
+    } else if (name === "Flight") {
+      effects.push(...this.activityMap[name].map((data) => {
+        return {
+          name: `Experimental Elixir: ${name}`,
+          activityMatch: `Use ${name}`,
+          type: "item",
+          options: {
+            transfer: false,
+          },
+          changes: [
+            DDBEnricherData.ChangeHelper.addChange(data.bonus, 20, "system.attributes.movement.fly"),
+          ],
+          data: {
+            "_id": utils.namedIDStub(name, {
+              postfix: data.bonus,
+              prefix: "ef",
+            }),
+            "flags.ddbimporter.effectIdLevel": {
+              min: data.min,
+              max: data.max,
+            },
+          },
+        };
+      }));
+    }
+
+    return effects;
+  }
+
+  _elixirEffects = null;
+
+  get getElixirEffects() {
+    if (this._elixirEffects) return this._elixirEffects;
+
+    const results = [];
+    for (const row of this.experimentalElixirDetails) {
+      const effects = this.generateElixirEffect(row.name);
+      results.push(...effects);
+    }
+    this._elixirEffects = results;
+    return results;
+
+  }
+
+  getElixirAdditionalEnchantActivityEffects(name) {
+    const results = [];
+    this.activityMap[name].forEach((m, i) => {
+      // const activityRiderIds = [];
+      // m.forEach((data, i) => {
+      //   activityRiderIds.push(utils.namedIDStub(name, {
+      //     postfix: i,
+      //     prefix: "eea",
+      //   }));
+      // });
+      results.push({
+        name,
+        type: "enchant",
+        changes: this.enchantChangeMap(name),
+        data: {
+          flags: {
+            ddbimporter: {
+              activityRiders: [utils.namedIDStub(name, {
+                postfix: i,
+                prefix: "eea",
+              })],
+              effectRiders: [],
+            },
+          },
+        },
+      });
     });
-    basicActivityGenerator.build(options);
 
-    // TODO generate and add effects
-    if (row.name === "Swiftness") {
+    return results;
+  }
 
-    } else if (row.name === "Resilience") {
-
-    } else if (row.name === "Boldness") {
-
-    } else if (row.name === "Flight") {
-
+  get elixirAdditionalActivities() {
+    const results = [];
+    for (const row of this.experimentalElixirDetails) {
+      const activities = this.generateElixirAdditionalActivity(row.name);
+      results.push(...activities);
     }
+    return results;
   }
 
-  async importItem(item) {
+  get additionalActivities() {
+    const base = [
+      {
+        constructor: {
+          name: "Crate Elixir",
+          type: "enchant",
+        },
+        overrides: {
+          activationType: "action",
+          addItemConsume: true,
+          targetType: "self",
+          data: {
+            restrictions: {
+              type: "consumable",
+            },
+          },
+        },
+      },
+      {
+        constructor: {
+          name: "Crate Elixir With Spell Slot",
+          type: "enchant",
+        },
+        overrides: {
+          activationType: "action",
+          addItemConsume: true,
+          targetType: "self",
+          data: {
+            restrictions: {
+              type: "consumable",
+            },
+            consumption: {
+              scaling: { allowed: true, max: "9" },
+              targets: [
+                {
+                  type: "spellSlots",
+                  value: "1",
+                  target: "1",
+                  scaling: { allowed: false, max: "" },
+                },
+              ],
+            },
+          },
+        },
+      },
+    ];
+
+    return base.concat(this.elixirAdditionalActivities);
+  }
+
+
+  get elixirEnchantEffects() {
+    const results = [];
+    for (const row of this.experimentalElixirDetails) {
+      const effect = {
+        name: `Experimental Elixir: ${row.name}`,
+        type: "enchant",
+      };
+      results.push(effect);
+    }
+    return results;
+  }
+
+  get effects() {
+    const baseEffects = [];
+
+    baseEffects.push(...this.getElixirEffects);
+
+    for (const row of this.experimentalElixirDetails) {
+      const enchant = this.getElixirAdditionalEnchantActivityEffects(row.name);
+      baseEffects.push(...enchant);
+    }
+
+    return baseEffects;
+  }
+
+  async buildItem(row) {
+    const itemData = this.getSkeletonItem(row);
+    // itemData.system.activities = this.elixirActivities[row.name] ?? [];
+    // itemData.effects = this.elixirEffects[row.name] ?? [];
+
+    return itemData;
+  }
+
+  elixirs = [];
+
+  static featureHandlerOptions = {
+    chrisPremades: true,
+    removeSRDDuplicates: false,
+    filterDuplicates: false,
+    deleteBeforeUpdate: false,
+    matchFlags: ["is2014", "is2024"],
+    useCompendiumFolders: true,
+    indexFilter: {
+      fields: [
+        "name",
+        "flags.ddbimporter",
+        "system.type.subtype",
+      ],
+    },
+  };
+
+  async importElixirs() {
+    const updateFeatures = this.updateCompendiumItems
+      ?? this.ddbParser.ddbCharacter.forceCompendiumUpdate
+      ?? game.settings.get(SETTINGS.MODULE_ID, "character-update-policy-update-add-features-to-compendiums");
+
+    const featureHandler = await DDBItemImporter.buildHandler("features", this.elixirs, updateFeatures, ExperimentalElixir.featureHandlerOptions, this.handler);
+    await featureHandler.buildIndex(ExperimentalElixir.featureHandlerOptions.indexFilter);
 
   }
 
-  async customFunction(options = {}) {
+  async buildCompendiumFolders() {
+    this.compendiumFolders = new DDBCompendiumFolders("features");
+    await this.compendiumFolders.loadCompendium("features");
+    const versionStub = this.is2014 ? "2014" : "2024";
+    await this.compendiumFolders.createSubClassFeatureFolder("Alchemist", "Artificer", versionStub);
+  }
+
+  async generateElixirs() {
+    await this.buildCompendiumFolders();
+
     for (const row of this.experimentalElixirDetails) {
       const item = await this.buildItem(row);
-      await this.importItem(item);
+      this.elixirs.push(item);
     }
+
+    await this.compendiumFolders.addCompendiumFolderIds(this.elixirs);
+    await this.importElixirs();
+  }
+
+  updateDescriptionTable(updates = []) {
+    const doc = utils.htmlToDoc(this.ddbParser?.ddbDefinition?.description ?? "");
+
+    const rows = doc.body.querySelectorAll('tbody tr');
+
+    for (const update of updates) {
+      const currentTitle = `${update.name.trim()}.`;
+
+      for (const row of rows) {
+        const cells = row.querySelectorAll('td');
+        if (cells.length < 2)
+          continue;
+
+        const effectCell = cells[1];
+        const strongTag = effectCell.querySelector('strong');
+
+        if (strongTag && strongTag.textContent.trim() === currentTitle) {
+          strongTag.textContent = `@UUID[${update.uuid}]{${update.name}.}`;
+          break;
+        }
+      }
+    }
+    foundry.utils.setProperty(this.ddbParser, "ddbDefinition.description", doc.body.innerHTML);
+  }
+
+  linkUpItemUUIDs() {
+    const updates = [];
+    for (const elixir of this.elixirs) {
+      const uuid = this.handler.compendiumIndex.find((e) => e._id === elixir._id)?.uuid
+        ?? this.handler.compendiumIndex.find((e) => e.name === elixir.name && e.flags?.ddbimporter?.is2014 === elixir.flags?.ddbimporter?.is2014)?.uuid;
+      if (!uuid) continue;
+      updates.push({ name: elixir.name.split(':').pop().trim(), uuid });
+    }
+
+    this.updateDescriptionTable(updates);
+
+  }
+
+  async customFunction() {
+    this.handler = new DDBItemImporter("features", [], ExperimentalElixir.featureHandlerOptions);
+    if (game.user.isGM) await this.generateElixirs();
+    this.linkUpItemUUIDs();
   }
 
 }